feat(db): initialize schema for libraries and audiobooks
- Added `init_schema` function to create fundamental tables for libraries and audiobooks in the database.
- Created indexes for efficient querying on audiobooks.
- Introduced a new module `schema.rs` for schema-related operations.

refactor(error): reorganize error handling macros

- Rearranged the order of modules in `error.rs` for better clarity.
- Enhanced documentation for error handling macros in `macros.rs` and `conversions.rs`.
- Improved organization of error conversion macros for better maintainability.

style(gui): improve code formatting in table_core.rs

- Adjusted indentation and formatting for better readability in `table_core.rs`.

fix(image_cache): simplify image data handling

- Refactored image data handling logic in `image_cache.rs` to use pattern matching for cleaner code.

docs(refactoring): outline database module refactoring plan

- Created a detailed refactoring plan for the `db/mod.rs` file to reduce its size and improve organization.
- Proposed splitting the module into focused submodules for better maintainability and clarity.
