+++ abop-core/src/error.rs
@@ -5,14 +5,15 @@ use hound;
 use thiserror::Error;
 use toml;
 
-pub mod macros;
-pub mod context;
 pub mod chain;
+pub mod context;
+pub mod macros;
 
 // Re-export commonly used macros and utilities
-pub use macros::*;
-pub use context::ErrorContext;
 pub use chain::ErrorChain;
+pub use context::ErrorContext;
+#[allow(unused_imports)]
+pub use macros::*;
 
 /// Central error type for the ABOP application
 ///
