+++ abop-core/src/error/macros/constructors.rs
@@ -1,6 +1,15 @@
-// Macros for constructing AppError variants and simple guards
+//! Constructors for common `AppError` variants and simple guards.
+//!
+//! These macros provide concise ways to create strongly-typed application
+//! errors throughout the codebase. Prefer these over ad‑hoc strings to keep
+//! error handling consistent.
 
 #[macro_export]
+/// Create a `AppError::Config` from a string or format args.
+///
+/// Examples
+/// - `config_error!("invalid setting")`
+/// - `config_error!("invalid setting: {}", key)`
 macro_rules! config_error {
     ($msg:expr) => {
         $crate::error::AppError::Config($msg.to_string())
@@ -11,6 +20,11 @@ macro_rules! config_error {
 }
 
 #[macro_export]
+/// Create a `AppError::Audio` from a string or format args.
+///
+/// Examples
+/// - `audio_error!("decoder failed")`
+/// - `audio_error!("decoder {} failed", name)`
 macro_rules! audio_error {
     ($msg:expr) => {
         $crate::error::AppError::Audio($msg.to_string())
@@ -21,6 +35,11 @@ macro_rules! audio_error {
 }
 
 #[macro_export]
+/// Create a `AppError::Database` with an `ExecutionFailed` variant.
+///
+/// Examples
+/// - `database_error!("unique constraint violated")`
+/// - `database_error!("query failed: {}", sql)`
 macro_rules! database_error {
     ($msg:expr) => {
         $crate::error::AppError::Database($crate::db::error::DatabaseError::ExecutionFailed {
@@ -35,6 +54,11 @@ macro_rules! database_error {
 }
 
 #[macro_export]
+/// Create a `AppError::ValidationFailed` from a string or format args.
+///
+/// Examples
+/// - `validation_error!("name cannot be empty")`
+/// - `validation_error!("{} must be >= {}", field, min)`
 macro_rules! validation_error {
     ($msg:expr) => {
         $crate::error::AppError::ValidationFailed($msg.to_string())
@@ -45,6 +69,11 @@ macro_rules! validation_error {
 }
 
 #[macro_export]
+/// Create a `AppError::InvalidData` from a string or format args.
+///
+/// Examples
+/// - `invalid_data!("unexpected file format")`
+/// - `invalid_data!("bad header: {}", header)`
 macro_rules! invalid_data {
     ($msg:expr) => {
         $crate::error::AppError::InvalidData($msg.to_string())
@@ -55,6 +84,9 @@ macro_rules! invalid_data {
 }
 
 #[macro_export]
+/// Create a `AppError::Timeout` capturing operation and timing details.
+///
+/// Example: `timeout_error!("scan library", 5_000, elapsed_ms)`
 macro_rules! timeout_error {
     ($operation:expr, $timeout_ms:expr, $elapsed_ms:expr) => {
         $crate::error::AppError::Timeout {
